Refactor cdp to share code with other browsers
Use `chrome-launcher` to run chrome
Remove some `any's`

Fix #311
Fix #330
Fix #324
Fix #332
