+++ package.json
@@ -17,6 +17,7 @@
     "caniuse-api": "^2.0.0",
     "canvas-prebuilt": "^1.6.0-canvas.5",
     "chalk": "^1.1.3",
+    "chrome-launcher": "^0.2.1",
     "chrome-remote-interface": "^0.23.2",
     "content-type": "^1.0.2",
     "debug": "^2.6.8",
