+++ src/lib/utils/location-helpers.ts
@@ -3,6 +3,8 @@ import { IAsyncHTMLAttribute, IAsyncHTMLElement, IProblemLocation } from './../t
 
 const debug: debug.IDebugger = d(__filename);
 
+const doubleQuotesRegex: RegExp = /"/g;
+
 /**
  * Creates a CSS selector from a given element using its attributes and the type of node:
  *
@@ -17,7 +19,7 @@ const selectorFromElement = (element: IAsyncHTMLElement): string => {
     for (let i = 0; i < attributes.length; i++) {
         const attribute: IAsyncHTMLAttribute | NamedNodeMap = attributes[i];
 
-        selector += `[${attribute.name}="${attribute.value}"]`;
+        selector += `[${attribute.name}="${attribute.value.replace(doubleQuotesRegex, '\\"')}"]`;
     }
 
     debug(`Selector created: ${selector}`);
